Initial commit
This is the result of asking Cursor to create a gradle plugin given the
prompt "Help me create a gradle plugin for Happo here. There should be two actions:

`./gradlew createHappoReport` -- take all screenshots from a specified folder and upload to happo.
`./gradlew compareHappoReports <sha1> <sha2>` -- compare two reports as identified by sha1 and sha2
"

There was a lot of back-and-forth but in the end things were building
correctly. I'll spend time to polish the generated code next.